+++ lib/jsdom/living/nodes/Document-impl.js
@@ -12,6 +12,7 @@ const whatwgURL = require("whatwg-url");
 const StyleSheetList = require("../generated/StyleSheetList.js");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const eventAccessors = require("../helpers/create-event-accessor");
+const { querySelector, querySelectorAll } = require("../helpers/selector");
 const { asciiLowercase, stripAndCollapseASCIIWhitespace } = require("../helpers/strings");
 const { childTextContent } = require("../helpers/text");
 const { HTML_NS, SVG_NS } = require("../helpers/namespaces");
@@ -87,6 +88,11 @@ function toLastModifiedString(date) {
     " " + pad(date.getHours()) +
     ":" + pad(date.getMinutes()) +
     ":" + pad(date.getSeconds());
+}
+
+function shouldAlwaysSelectNothing(elImpl) {
+  // This is true during initialization.
+  return elImpl === elImpl._ownerDocument && !elImpl.documentElement;
 }
 
 const eventInterfaceTable = {
@@ -909,6 +915,22 @@ class DocumentImpl extends NodeImpl {
     copy._URL = this._URL;
     copy._origin = this._origin;
     return copy;
+  }
+
+  querySelector(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return null;
+    }
+    return idlUtils.implForWrapper(querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject));
+  }
+
+  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
+  querySelectorAll(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return NodeList.create(this._globalObject, [], { nodes: [] });
+    }
+    const list = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
+    return NodeList.create(this._globalObject, [], { nodes: list.map(n => idlUtils.tryImplForWrapper(n)) });
   }
 }
