+++ lib/jsdom/living/nodes/DocumentFragment-impl.js
@@ -1,11 +1,18 @@
 "use strict";
 const { mixin } = require("../../utils");
 const { domSymbolTree } = require("../helpers/internal-constants");
+const { querySelector, querySelectorAll } = require("../helpers/selector");
 const NODE_TYPE = require("../node-type");
 const NodeImpl = require("./Node-impl").implementation;
 const NonElementParentNodeImpl = require("./NonElementParentNode-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
 const idlUtils = require("../generated/utils");
+const NodeList = require("../generated/NodeList");
+
+function shouldAlwaysSelectNothing(elImpl) {
+  // This is true during initialization.
+  return elImpl === elImpl._ownerDocument && !elImpl.documentElement;
+}
 
 class DocumentFragmentImpl extends NodeImpl {
   constructor(globalObject, args, privateData) {
@@ -33,6 +40,22 @@ class DocumentFragmentImpl extends NodeImpl {
     }
 
     return null;
+  }
+
+  querySelector(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return null;
+    }
+    return idlUtils.implForWrapper(querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject));
+  }
+
+  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
+  querySelectorAll(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return NodeList.create(this._globalObject, [], { nodes: [] });
+    }
+    const list = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
+    return NodeList.create(this._globalObject, [], { nodes: list.map(n => idlUtils.tryImplForWrapper(n)) });
   }
 }
