+++ lib/jsdom/living/nodes/Element-impl.js
@@ -11,10 +11,12 @@ const { parseFragment } = require("../../browser/parser");
 const InnerHTMLImpl = require("../domparsing/InnerHTML-impl").implementation;
 const { fragmentSerialization } = require("../domparsing/serialization");
 const { domSymbolTree } = require("../helpers/internal-constants");
-const { closest, matches } = require("../helpers/selector");
+const { closest, matches, querySelector, querySelectorAll } = require("../helpers/selector");
 const DOMException = require("domexception/webidl2js-wrapper");
 const DOMTokenList = require("../generated/DOMTokenList");
 const NamedNodeMap = require("../generated/NamedNodeMap");
+const NodeList = require("../generated/NodeList");
+const idlUtils = require("../generated/utils");
 const validateNames = require("../helpers/validate-names");
 const { asciiLowercase, asciiUppercase } = require("../helpers/strings");
 const { listOfElementsWithQualifiedName, listOfElementsWithNamespaceAndLocalName,
@@ -50,6 +52,11 @@ function detachId(id, elm, doc) {
       }
     }
   }
+}
+
+function shouldAlwaysSelectNothing(elImpl) {
+  // This is true during initialization.
+  return elImpl === elImpl._ownerDocument && !elImpl.documentElement;
 }
 
 class ElementImpl extends NodeImpl {
@@ -550,6 +557,22 @@ class ElementImpl extends NodeImpl {
 
   webkitMatchesSelector(selectors) {
     return this.matches(selectors);
+  }
+
+  querySelector(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return null;
+    }
+    return idlUtils.implForWrapper(querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject));
+  }
+
+  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
+  querySelectorAll(selectors) {
+    if (shouldAlwaysSelectNothing(this)) {
+      return NodeList.create(this._globalObject, [], { nodes: [] });
+    }
+    const list = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
+    return NodeList.create(this._globalObject, [], { nodes: list.map(n => idlUtils.tryImplForWrapper(n)) });
   }
 }
