+++ lib/jsdom/living/nodes/ParentNode-impl.js
@@ -1,10 +1,7 @@
 "use strict";
 
-const idlUtils = require("../generated/utils");
-const NodeList = require("../generated/NodeList");
 const HTMLCollection = require("../generated/HTMLCollection");
 const { domSymbolTree } = require("../helpers/internal-constants");
-const { querySelector, querySelectorAll } = require("../helpers/selector");
 const NODE_TYPE = require("../node-type");
 const { convertNodesIntoNode } = require("../node");
 
@@ -60,27 +57,6 @@ class ParentNodeImpl {
     this._preInsertValidity(node, null);
     this._replaceAll(node);
   }
-
-  querySelector(selectors) {
-    if (shouldAlwaysSelectNothing(this)) {
-      return null;
-    }
-    return querySelector(selectors, this, this._globalObject);
-  }
-
-  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
-  querySelectorAll(selectors) {
-    if (shouldAlwaysSelectNothing(this)) {
-      return NodeList.create(this._globalObject, [], { nodes: [] });
-    }
-    const list = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-    return NodeList.create(this._globalObject, [], { nodes: list.map(n => idlUtils.tryImplForWrapper(n)) });
-  }
-}
-
-function shouldAlwaysSelectNothing(elImpl) {
-  // This is true during initialization.
-  return elImpl === elImpl._ownerDocument && !elImpl.documentElement;
 }
 
 module.exports = {
